Split argument parser tests, add new tests, polish old ones.
