+++ methods/__init__.py
@@ -1,2 +1,1 @@
-
 from .gsea import *
